Move generic game properties fomr BaseballGame to Game class
