+++ src/app/espn/mlb/class/baseballGame.class.ts
@@ -10,38 +10,13 @@ import {
 } from '../interface/game';
 import { MLB_STADIUM_MAP, MLB_TEAM } from '../consts/team.const';
 import { domeStadiums } from '../mlb.const';
+import { Game } from '@app/@shared/class';
 
-export class BaseballGame {
-  private _event: EspnClientEvent;
-  private _competitors: { [id: number]: EspnClientCompetitor } = {};
-  private _homeTeam: number;
-  private _awayTeam: number;
+export class BaseballGame extends Game {
   private _currentConditions: CurrentConditions;
-
-  constructor(event: EspnClientEvent) {
-    this._event = event;
-  }
-
-  get gameId() {
-    return this._event.id;
-  }
 
   set currentConditions(val: CurrentConditions) {
     this._currentConditions = val;
-  }
-
-  set competitors(val: EspnClientCompetitor[]) {
-    const map: { [id: number]: EspnClientCompetitor } = {};
-    for (const comp of val) {
-      map[Number(comp.id)] = comp;
-
-      if (comp.homeAway === 'home') {
-        this._homeTeam = Number(comp.id);
-      } else {
-        this._awayTeam = Number(comp.id);
-      }
-    }
-    this._competitors = map;
   }
 
   get homeTeam(): ScoreboardGameTeam {
@@ -68,20 +43,6 @@ export class BaseballGame {
     return {
       event: this._event.summary,
       game: `${MLB_TEAM[this._awayTeam]} at ${MLB_TEAM[this._homeTeam]}`,
-    };
-  }
-
-  get gameDate(): ScoreboardGameStart {
-    return {
-      iso: this._event.date,
-      milli: new Date(this._event.date).getTime(),
-    };
-  }
-
-  get status(): ScoreboardGameStatus {
-    return {
-      isFinal: this._event.fullStatus.type.completed || this._event.fullStatus.type.name === 'STATUS_POSTPONED',
-      completePerc: this._event.percentComplete,
     };
   }
